+++ common/components/News/ContentForPhoto.js
@@ -35,8 +35,16 @@ class ContentForPhoto extends PureComponent {
     };
     const randomKey = news.sn % 3;
     const imgApi = 'https://imgapiv2.nownews.com/?w=1080&q=70&src=';
+    let imgUrl = '';
+
+    if (MainPhoto && MainPhoto.googleCDN) {
+      imgUrl = MainPhoto.googleCDN;
+    } else if (MainPhoto && MainPhoto.url) {
+      imgUrl = `${imgApi}${MainPhoto.url}`;
+    }
+
     const socialProps = {
-      img: MainPhoto && MainPhoto.url,
+      img: imgUrl,
       title: news.title,
       url: news.completeUrl
     };
@@ -44,9 +52,9 @@ class ContentForPhoto extends PureComponent {
       <Container>
         <div className={css(styles.SlideBox)}>
           <Carousel {...settings}>
-            {Photos.map(({ desc, sn, url }) => (
+            {Photos.map(({ desc, sn, googleCDN, url }) => (
               <div className={css(styles.contentDiv)} key={sn}>
-                <img className={css(styles.contentImg)} src={`${imgApi}${url}`} />
+                <img className={css(styles.contentImg)} src={googleCDN || url} />
                 <p className={css(styles.imgDesc)}>{desc}</p>
               </div>
             ))}