+++ common/routes/Video/components/BlockItems9.js
@@ -24,7 +24,7 @@ class BlockItems9 extends PureComponent {
       return (
         <div key={sn} className={css(styles.blockItem)} onClick={() => { this.select(index); }}>
           <TypeIcon type='VIDEO' />
-          <img className={css(styles.img)} src={photo.thumbnail || photo.url} />
+          <img className={css(styles.img)} src={photo.googleCDN || photo.thumbnail || photo.url} />
           <div className={css(styles.bottom)}>
             <div className={css(styles.title)}><h3 className={css(styles.h3)}>{shortTitle}</h3></div>
             <img src='/icons/whiteClock.png' />